Moved the login helpers to a separate file under support, added log_me_in in the features in the before block
