+++ app/views/ideas/index.html.slim
@@ -13,12 +13,10 @@ javascript:
   $('.upvote img').click( function() {
     $el = $(this);
     url = "/ideas/" + $el.data('idea-id') + "/upvote";
-    console.log(url);
     $.ajax({
       url: url,
       type: 'put'
     }).done(function(data) {
-      console.log(data.votes);
       $el.parent().parent().find('.votes').text(data.votes);
     });
   });
