+++ spec/features/upvote_idea_spec.rb
@@ -3,6 +3,10 @@ require 'features/features_helper'
 feature "Upvoting an idea increases the votes", :js do
   let!(:ideas) { FactoryGirl.create_list(:idea, 10) }
 
+  before do
+    log_me_in
+  end
+  
   scenario "Clicking on an upvote button" do
     visit "/"
